pyproject.toml
[project]
name = "majaai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "appdirs==1.4.4",
    "asttokens==3.0.0",
    "attrs==25.4.0",
    "backcall==0.2.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.2",
    "black==25.1.0",
    "bleach==6.3.0",
    "blinker==1.9.0",
    "boto3==1.40.66",
    "botocore==1.40.66",
    "build==1.3.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.4",
    "chromadb==1.3.5",
    "click==8.3.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "datasets==4.4.1",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distlib==0.4.0",
    "distro==1.9.0",
    "docopt==0.6.2",
    "docstring-parser==0.17.0",
    "durationpy==0.10",
    "executing==2.2.1",
    "faiss-cpu==1.12.0",
    "fastapi>=0.121.3",
    "fastjsonschema==2.21.2",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "flatbuffers==25.9.23",
    "fonttools==4.60.0",
    "frozenlist==1.7.0",
    "fsspec==2025.10.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-ai-generativelanguage==0.7.0",
    "google-api-core==2.25.1",
    "google-auth==2.43.0",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.2.4",
    "grpcio==1.76.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==1.1.4",
    "humanfriendly==10.0",
    "identify==2.6.15",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "instructor==1.12.0",
    "ipykernel==6.30.1",
    "ipython==8.12.3",
    "ipython-pygments-lexers==1.1.1",
    "isort==6.0.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpath-ng==1.7.0",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyterlab-pygments==0.3.0",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-aws==0.2.31",
    "langchain-community==0.3.29",
    "langchain-core==0.3.79",
    "langchain-redis==0.2.4",
    "langchain-text-splitters==0.3.11",
    "langchainhub==0.1.21",
    "langgraph==0.6.7",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.7",
    "langsmith==0.4.28",
    "langwatch>=0.6.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.1.4",
    "ml-dtypes==0.5.3",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "multiprocess==0.70.18",
    "mypy-extensions==1.1.0",
    "narwhals==2.5.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "nltk==3.9.2",
    "nodeenv==1.9.1",
    "numpy==2.3.5",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.2",
    "openai==1.107.3",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.4",
    "ormsgpack==1.10.0",
    "overrides==7.7.0",
    "pandas==2.3.2",
    "pandocfilters==1.5.1",
    "parso==0.8.5",
    "pathspec==0.12.1",
    "pickleshare==0.7.5",
    "pillow==11.3.0",
    "pinecone==7.3.0",
    "pinecone-plugin-assistant==1.8.0",
    "pinecone-plugin-interface==0.0.7",
    "pipreqs==0.5.0",
    "platformdirs==4.4.0",
    "ply==3.11",
    "posthog==5.4.0",
    "pre-commit==4.4.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.33.1",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.10.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pymupdf==1.26.4",
    "pyparsing==3.2.4",
    "pypdf==6.0.0",
    "pypdf2==3.0.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-ulid==3.1.0",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "ragas==0.3.9",
    "rank-bm25==0.2.2",
    "redis==6.4.0",
    "redisvl==0.11.0",
    "referencing==0.37.0",
    "regex==2025.9.1",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rpds-py==0.29.0",
    "rsa==4.9.1",
    "s3transfer==0.14.0",
    "scikit-network==0.33.3",
    "scipy==1.16.3",
    "shellingham==1.5.4",
    "simsimd==6.5.3",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.43",
    "stack-data==0.6.3",
    "streamlit==1.49.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tiktoken==0.11.0",
    "tinycss2==1.4.0",
    "tokenizers==0.22.1",
    "toml==0.10.2",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.20.0",
    "typer-slim==0.20.0",
    "types-requests==2.32.4.20250913",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3>=2.4.0",
    "uvicorn==0.38.0",
    "virtualenv==20.35.4",
    "watchdog==6.0.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "yarg==0.1.9",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
